Fix fuzzer timeouts in anim_util_fuzzer and imageio_fuzzer.
- In anim_util_fuzzer, check for the 3-byte 'GIF' prefix rather
than only 'GIF87a'/'GIF89a' so 'GIFVER' inputs (accepted by IsGIF()
and DGifOpen()) do not bypass dimension and frame-count checks,
and return early on malformed GIF blocks instead of breaking out
into ReadAnimatedImage(). Also bound total WebP animation frames
to kFuzzFrameLimit.
- In fuzz_utils, update GetSOFData() to handle standalone JPEG
markers (RST0-RST7, SOI, TEM), fill bytes, and all JPEG SOF
markers (SOF0-SOF15), and extend IsImageTooBig() to check PNG
(IHDR) and PNM header dimensions.
- In imageio_fuzzer, check TIFF dimensions and tile sizes via
TIFFClientOpen before invoking ReadTIFF().

+++ tests/fuzzer/fuzz_utils.h
@@ -49,13 +49,13 @@ namespace fuzz_utils {
 // Enabling some sanitizers slow down runtime significantly.
 // Use a very low threshold in this case to avoid timeouts.
 #if defined(__SANITIZE_ADDRESS__)  // GCC
-static const size_t kFuzzPxLimit = 1024 * 1024 / 10;
+static const uint64_t kFuzzPxLimit = 1024 * 1024 / 10;
 #elif !defined(__has_feature)  // Clang
-static const size_t kFuzzPxLimit = 1024 * 1024;
+static const uint64_t kFuzzPxLimit = 1024 * 1024;
 #elif __has_feature(address_sanitizer) || __has_feature(memory_sanitizer)
-static const size_t kFuzzPxLimit = 1024 * 1024 / 18;
+static const uint64_t kFuzzPxLimit = 1024 * 1024 / 18;
 #else
-static const size_t kFuzzPxLimit = 1024 * 1024;
+static const uint64_t kFuzzPxLimit = 1024 * 1024;
 #endif
 
 // Demuxed or decoded animation frames.
